% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescargarModis.R
\name{DescargarModis}
\alias{DescargarModis}
\title{Descargar producto Modis}
\usage{
DescargarModis(Coleccion, Producto, Salida, InicioFecha, FinFecha, LngMin,
  LatMin, LngMax, LatMax)
}
\arguments{
\item{Producto}{Corresponde al producto (char)}

\item{Salida}{Corresponde al direcorio de salida (char)}

\item{InicioFecha}{Corresponde a la fecha inicial (char) <2001-01-01>}

\item{FinFecha}{Corresponde a la fecha final (char) <2001-12-31>}

\item{LngMin}{Corresponde al valor mínimo de longitud (char,double) <Coordendas en WGS84-4326>}

\item{LatMin}{Corresponde al valor mínimo de latitud (char,double) <Coordendas en WGS84-4326>}

\item{LngMax}{Corresponde al valor máximo de longitud (char,double) <Coordendas en WGS84-4326>}

\item{LatMax}{Corresponde al valor máximo de latitud (char,double) <Coordendas en WGS84-4326>}

\item{coleccion}{Corresponde a la colección del producto (int)}
}
\value{
Retorna los archivos almacenados en la ruta de salida
}
\description{
Este comando permite realizar la descarga de los productos Modis
}
\examples{
DescargarModis(5,"MOD03","C:/DOWNLOAD/MODIS/MOD03/","2001-01-01","2001-12-31",-81.7,-4.2,-66.8,13.3)

coleccion = ListarColeccionModis()[8]
producto = ListarProductoModis(coleccion)[213]
salida = "c:/DOWNLOAD/MODIS/MOD03"
fecha_o = "2001-01-01"
fecha_f = "2001-12-31"
lngmin = -81.7
latmin = -4.2
lngmax = -66.8
latmax = 13.3
DescargarModis(coleccion,producto,salida,fecha_o,fecha_f,lngmin,latmin,lngmax,latmax)

}
