% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListarAniosModis.R
\name{ListarAniosModis}
\alias{ListarAniosModis}
\title{Listar anios del producto Modis}
\usage{
ListarAniosModis(coleccion, producto)
}
\arguments{
\item{coleccion}{Corresponde a la colección del producto (int)}

\item{producto}{Corresponde al producto de Modis (char)}
}
\value{
Retorna el vector del listado de anios (list)
}
\description{
Este comando permite retornar el listado de anios existentes en el producto Modis seleccionado
}
\examples{
ListarAniosModis(5,"MOD03") #list()

coleccion = ListarColeccionModis()
producto = ListarProductoModis(coleccion[8])
ListarAniosModis(coleccion[8],producto[213]) #list()
}
