% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListarColeccionModis.R
\name{ListarColeccionModis}
\alias{ListarColeccionModis}
\title{Listar Coleccion de Modis}
\usage{
ListarColeccionModis()
}
\value{
Retorna el vector del listado de colecciones (list)
}
\description{
Este comando permite retornar el listado de las colecciones
}
\examples{
coleccion = ListarColeccionModis() #list()
}
