% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListarAniosModis.R
\name{ListarAniosModis}
\alias{ListarAniosModis}
\title{Listar a침os del producto Modis}
\usage{
ListarAniosModis(coleccion, producto)
}
\arguments{
\item{coleccion}{Corresponde a la colecci칩n del producto (int)}

\item{producto}{Corresponde al producto de Modis (char)}
}
\value{
Retorna el vector del listado de a침os (list)
}
\description{
Este comando permite retornar el listado de a침os existentes en el producto Modis seleccionado
}
\examples{
ListarAniosModis(5,"MOD03") #list()

coleccion = ListarColeccionModis()[8]
producto = ListarProductoModis(coleccion)[213]
ListarAniosModis(coleccion,producto) #list()
}
